requirements.txt
streamlit == 1.35.0
numpy == 1.23.5
pandas == 1.4.4
pillow == 10.3.0
plotly == 5.22.0
altair == 5.3.0
wordcloud == 1.9.3
streamlit-option-menu == 0.3.13
keras == 2.15.0
regex == 2024.5.15
tensorflow == 2.15.0
stqdm == 0.0.5
wordcloud == 1.9.3
numerize == 0.12